Dev tilemap editinig feature

+++ src/client/TilemapSection.ts
@@ -1,7 +1,8 @@
 import { DomNode, el, Input, Store } from "@common-module/app";
-import { Screen, TilemapData } from "@gaiaengine/2d";
+import { Screen, Tilemap, TilemapData } from "@gaiaengine/2d";
+import { Graphics } from "pixi.js";
+import HoverTileDisplay from "../node/HoverTileDisplay.js";
 import InfiniteGrid from "../node/InfiniteGrid.js";
-import TileAreaDisplay from "../node/TileAreaDisplay.js";
 import EditorService from "./EditorService.js";
 
 export default class TilemapSection extends DomNode {
@@ -13,6 +14,8 @@ export default class TilemapSection extends DomNode {
   private dragging = false;
   private dragX = 0;
   private dragY = 0;
+  private touchstartX = 0;
+  private touchstartY = 0;
 
   private tileSizeInput: Input;
   private screen: Screen;
@@ -20,10 +23,19 @@ export default class TilemapSection extends DomNode {
   private yInput: Input;
   private zoomInput: Input;
 
+  private tilemap: Tilemap;
   private grid: InfiniteGrid;
-  private hoverTile: TileAreaDisplay | undefined;
+  private hoverTile: HoverTileDisplay | undefined;
 
-  constructor(private projectId: string, private tilemapData: TilemapData) {
+  private selectedTile:
+    | { key: string; row: number; col: number }
+    | undefined;
+
+  constructor(
+    private projectId: string,
+    private tilesets: { [key: string]: string },
+    private tilemapData: TilemapData,
+  ) {
     super("section.tilemap");
     this.addAllowedEvents("tileSizeChange");
 
@@ -32,12 +44,17 @@ export default class TilemapSection extends DomNode {
     this.y = this.transformStore.get("y") ?? 0;
     this.zoom = this.transformStore.get("zoom") ?? 1;
 
+    const tilemapTilesets: { [key: string]: string } = {};
+    for (const key in this.tilesets) {
+      tilemapTilesets[key] = `api/load-assets/${this.tilesets[key]}`;
+    }
+
     this.append(
       el(
         "header",
         this.tileSizeInput = new Input({
           label: "Tile Size",
-          value: tilemapData.tileSize.toString(),
+          value: this.tilemapData.tileSize.toString(),
         }),
       ),
       el(
@@ -45,7 +62,13 @@ export default class TilemapSection extends DomNode {
         this.screen = new Screen(
           0,
           0,
-          this.grid = new InfiniteGrid(0, 0, tilemapData.tileSize),
+          this.tilemap = new Tilemap(
+            0,
+            0,
+            tilemapTilesets,
+            this.tilemapData,
+          ),
+          this.grid = new InfiniteGrid(0, 0, this.tilemapData.tileSize),
         ),
       ),
       el(
@@ -61,9 +84,14 @@ export default class TilemapSection extends DomNode {
 
     this.screen.backgroundColor = 0xbfbfbf;
 
+    const centerGraphics = new Graphics();
+    centerGraphics.rect(-1, -1, 2, 2).fill(0xff0000);
+    centerGraphics.zIndex = 1000;
+    this.screen.root.container.addChild(centerGraphics);
+
     this.tileSizeInput.on("change", () => {
       const tileSize = parseInt(this.tileSizeInput.value);
-      tilemapData.tileSize = tileSize;
+      this.tilemapData.tileSize = tileSize;
       this.tileSizeInput.value = tileSize.toString();
       EditorService.saveTilemap(this.tilemapData);
 
@@ -78,6 +106,8 @@ export default class TilemapSection extends DomNode {
       this.dragging = true;
       this.dragX = event.clientX;
       this.dragY = event.clientY;
+      this.touchstartX = event.clientX;
+      this.touchstartY = event.clientY;
     });
 
     this.screen.onDom("mousemove", (event: MouseEvent) => {
@@ -95,7 +125,23 @@ export default class TilemapSection extends DomNode {
       this.touchMoveHandler(event);
     });
 
-    this.screen.onDom("mouseup", () => this.dragging = false);
+    this.screen.onDom("mouseup", (event: MouseEvent) => {
+      this.dragging = false;
+      if (
+        this.selectedTile &&
+        Math.abs(event.clientX - this.touchstartX) < 5 &&
+        Math.abs(event.clientY - this.touchstartY) < 5
+      ) {
+        const { row, col } = this.getRowColFromEvent(event);
+        this.tilemap.addTile({
+          row,
+          col,
+          tileset: this.selectedTile,
+        });
+        EditorService.saveTilemap(this.tilemapData);
+      }
+    });
+
     this.screen.onDom("touchmove", this.touchMoveHandler);
 
     this.screen.onDom("wheel", (event: WheelEvent) => {
@@ -130,7 +176,7 @@ export default class TilemapSection extends DomNode {
     this.onWindow("resize", () => this.resizeScreen());
   }
 
-  private touchMoveHandler = (e: TouchEvent | MouseEvent) => {
+  private getRowColFromEvent(e: TouchEvent | MouseEvent) {
     const screenRect = this.screen.rect;
     const rx = ((e instanceof TouchEvent ? e.touches[0].clientX : e.clientX) -
       screenRect.x - this.screen.width / 2 + this.screen.camera.x) /
@@ -138,25 +184,43 @@ export default class TilemapSection extends DomNode {
     const ry = ((e instanceof TouchEvent ? e.touches[0].clientY : e.clientY) -
       screenRect.y - this.screen.height / 2 + this.screen.camera.y) /
       this.screen.root.scaleY;
-
     const row = Math.floor(
       (ry + this.tilemapData.tileSize / 2) / this.tilemapData.tileSize,
     );
     const col = Math.floor(
       (rx + this.tilemapData.tileSize / 2) / this.tilemapData.tileSize,
     );
+    return { row, col };
+  }
 
+  private touchMoveHandler = (event: TouchEvent | MouseEvent) => {
+    const { row, col } = this.getRowColFromEvent(event);
     if (
-      this.hoverTile && this.hoverTile.row === row &&
-      this.hoverTile.col === col
+      !this.selectedTile || (
+        this.hoverTile && this.hoverTile.row === row &&
+        this.hoverTile.col === col
+      )
     ) return;
+
     this.hoverTile?.delete();
-    this.hoverTile = new TileAreaDisplay(this.tilemapData.tileSize, row, col)
-      .appendTo(this.screen.root);
+    this.hoverTile = new HoverTileDisplay(
+      this.tilemapData.tileSize,
+      row,
+      col,
+      this.tilemap.getTileTexture(
+        this.selectedTile.key,
+        this.selectedTile.row,
+        this.selectedTile.col,
+      ),
+    ).appendTo(this.screen.root);
   };
 
   private resizeScreen() {
     const rect = this.screen.parent!.rect;
     this.screen.resize(rect.width, rect.height);
+  }
+
+  public setTile(tilesetKey: string, row: number, col: number) {
+    this.selectedTile = { key: tilesetKey, row, col };
   }
 }
